refactor(split-layout): extract createSplitFocusTracker into its own module
Verbatim move of getParentSplitId and createSplitFocusTracker out of
SplitLayout.tsx (plus the export keyword) so the tracker can be
unit-tested without mounting the full split UI. No behavior change.
